Calculation of bonus
WAP to accept the current year and the year in which the employee joined the organization. If the number of years for which the employee has served the organization is greater than 3, then a bonus of $2500 is given to the employee. If the years of service are not greater than 3, then the program should do nothing.