Calculation of gross Salary
In a company an employee is paid as under: If his basic salary is less than $ 7500, then HRA = 10% of basic salary and DA = 90% of basic salary. If his salary is either equal to or above $7500, then HRA is $1000 and DA is 98% of basic salary. If the employee’s salary is input through the keyboard write a program to find his gross salary.