Check whether the triangle is scalene, Isosceles or equilateral
1.	Write a C program to accept the three sides of a triangle and check whether a triangle is Equilateral, Isosceles or Scalene. 
Test Data :
50 50 60
Expected Output :
This is an isosceles triangle.
Equilateral triangle: An equilateral triangle is a triangle in which all three sides are equal. In the familiar Euclidean geometry, equilateral triangles are also equiangular; that is, all three internal angles are also congruent to each other and are each 60°.
Isosceles triangle: An isosceles triangle is a triangle that has two sides of equal length.
Scalene triangle: A scalene triangle is a triangle that has three unequal sides.