Determine if the driver should be insured or not.
A company insures its drivers in the following cases:
a.	If the driver is married.
b.	If the driver is unmarried, male and above 30 years of age.
c.	If the driver is unmarried, female and above 25 years of age.
d.	In all other cases driver is not insured.
If the marital status, sex and age of the driver are the inputs, WAP to determine whether the driver should be insured or not. Use only “if” and “else” statements. Do not use Logical operators.