Check the grade of steel
1.	A certain grade of steel is graded according to the following conditions:
a.	Hardness must be greater than 50
b.	Carbon content must be less than 0.7
c.	Tensile strength must be greater than 5600
The grades are as follows:
-	Grade is 10 if all three conditions are met
-	Grade is 9 if conditions (a) and (b) are met
-	Grade is 8 if conditions (b) and (c) are met
-	Grade 7 if conditions (a) and (c) are met
-	Grade 6 if only one conditions is met
-	Grade 5 if none of the conditions are met
WAP, which will require the user to give values of hardness, carbon content and tensile strength of the steel under consideration and output the grade of the steel.